BreakModule Plugin
This plugin is intended to give the user the option to pause the
execution of the debugged  application when the specified modules are
loaded.
The modules specifications are saved to disk hence in any new debuggin
session the breakpoints are gonna remain untouched.
The user can include or exclude modules to the list as he or she wishes.